[ feat ] api 분리
